pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "okopilote-devices-usb-x440"
dynamic = ["version"]
description = ''
readme = "README.md"
requires-python = ">=3.7"
license = "MIT"
keywords = []
authors = [
  { name = "François", email = "francoismdj@gmx.com" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.7",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
  "pyserial",
  "okopilote.devices.common~=0.0.1",
]

[tool.hatch.build.targets.wheel]
packages = ["src/okopilote"]

[project.urls]
Documentation = "https://github.com/francoismdj/okopilote/devices/usb-x440#readme"
Issues = "https://github.com/francoismdj/okopilote/devices/usb-x440/issues"
Source = "https://github.com/francoismdj/okopilote/devices/usb-x440"

[tool.hatch.version]
path = "src/okopilote/devices/usb_x440/__about__.py"

[tool.hatch.envs.types]
extra-dependencies = [
  "mypy>=1.0.0",
]
[tool.hatch.envs.types.scripts]
check = "mypy --install-types --non-interactive {args:src/okopilote/devices/usb_x440 tests}"

[tool.coverage.run]
source_pkgs = ["okopilote", "tests"]
branch = true
parallel = true
omit = [
  "src/okopilote/devices/usb_x440/__about__.py",
]

[tool.coverage.paths]
okopilote_devices_usb_x440 = ["src/okopilote/devices/usb_x440", "*/usb-x440/src/okopilote/devices/usb_x440"]
tests = ["tests", "*/usb-x440/tests"]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]
